ci: Bump the timeout for building kernel+rootfs.
The build is absurdly long, and it's a waste to make it to an hour in and
get cancelled, only to have to try again and hope you hit a runner with a
hot ccache.  We'll still hit the 60min marge timeout sometimes, but at
least if the job passes eventually then we'll have the tarball up so that
next time it's a cache hit.