freedreno/drm: specify presumed buffer address in drm_msm_gem_submit_bo
This is a workaround that won't be required in the near future. See !27630.

Some kernels use the 'presumed' field in the drm_msm_gem_submit_bo struct
as the definitive buffer object address basis. When not set, any command
stream ends up in a GPU hang due to incorrect iova calculation in the
kernel.

Signed-off-by: Zan Dobersek <zdobersek@igalia.com>