ci/android: get prebuilt cuttlefish images from the net
To test mesa on cuttlefish special images tailored for upstream mesa are
needed, in particular these images set some env variables and some
properties so that surfaceflinger can reliably use upstream mesa.

So instead of downloading the standard images from Google, get those
images from the network.

NOTE: this is a temporary , the plan is to have artifacts stored in an
S3 bucket and let the ci scripts get it from there.