docs: rework ext role
It's only the GL-case that needs complicated logic here, because of
having to use use a path that contains the vendor-prefix twice. The
vk-ext role can just be a trivial extlink.

So let's split these in two, and keep things a bit simpler. We no
longer need the GL_/VK_-prefixes, so let's drop them to avoid making
the role-usage more verbose than needed.