ci/android: don't need debian-testing for the android test job
Since we are experimenting with the llvmpipe driver we don't need to
build an update mesa on the linux host, building for the Android host
only is enough.

Also keep the artifacts for the debian-android job to make the
after_script commands, which expect some artifacts, still succeed.

NOTE: this needs some cleanup as it might not be general enough: in
particular the .test common job still has an ldd invocation that assumes
mesa has been built on the host, and not cross-built for another
platform:
https://gitlab.freedesktop.org/mesa/mesa/-/blob/79637d5942d8933841f7d93d06b864e827d61aec/.gitlab-ci/test/gitlab-ci.yml#L11-13