ci/swrast: Add .test-in-situ to inherit
Right now all our hardware drivers inherit from common bases, which is
nice. Our test-in-place setups for llvmpipe/softpipe/virgl, however,
rely on the script being set by whichever of test-deqp or whatever that
they land on.

It would be good to get rid of that so we have a clearer path and fewer
instances of helper scripts randomly not being called, xvfb being run,
etc. Adding .test-in-situ is a first step there.

Signed-off-by: Daniel Stone <daniels@collabora.com>