etnaviv: nir: Enable fneg lowering
In the next commit, we will switch to using derivative intrinsics. Unlike
ALU operations, NIR intrinsics don't have a source swizzle, so a mov will
be added before each derivative intrinsic. Because of this, it's not
possible to apply source modifiers directly to derivative intrinsics in NIR.

It seems we're encountering some edge cases in the GPU ISA related
to source modifiers, which are causing a few Piglit failures. To
simplify things, let's lower fneg directly in NIR.

Signed-off-by: Christian Gmeiner <cgmeiner@igalia.com>