ci/android: add a manual job to build a AOSP cuttlefish image
Add a manual job to build a custom AOSP cuttlefish image and upload it
to the S3 bucket, this is useful because some customizations on the AOPS
side are needed to be able to have upstream mesa working in cuttlefish.

The custom image will be used for testing upstream mesa on Android.

Note that building AOSP is very expensive but this job is intended to be
run manually, and very rarely.