panfrost: fix SAMPLE_ALPHA_TO_ONE
When SAMPLE_ALPHA_TO_ONE is enabled we need to use a blend shader, as
the BlendDescriptor.alpha_to_one bit is deprecated (does not work)
on bifrost and valhall.

In order to make the blend shader set alpha to one, we also need to
pass the alpha_to_one status in to the blend shader creation
functions.

Signed-off-by: Eric R. Smith <eric.smith@collabora.com>