ci/android: add a manual job to build LLVM libraries for Android
In order to build the mesa llvmpipe driver for Android, LLVM libraries
for Android are needed, but building them each time a new upstream mesa
is built would be too much overhead.

So add a manual job to build LLVM libraries for Android and upload them
to the S3 bucket, the job will be called manually only when necessary
and the resulting prebuilt libraries will be re-used when building mesa
for Android.

The build process follows what has been documented in
docs/drivers/llvmpipe.rst

Reviewed-by: Guilherme Gallo <guilherme.gallo@collabora.com>