iris: Resolve compressed images to share by copying (xe2)
We need to resolve a compressed image that wans't created with
a modifier supporting compression prior to sharing it. On Xe2+
platforms, this is done by copying the memory to an uncompressed
location.

Fixed: dmabufshare
https://gitlab.freedesktop.org/mesa/demos/-/blob/main/src/egl/
opengl/dmabufshare.c?ref_type=heads

Image in client window gets misrendered because the driver at
client side received a compressed linear image.

Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>