mesa/st: create hash_table_u64 for drawable
Using address of drawable as hash table key will cause memory issue in this
situation:

1. drawable A with address addr is destroyed and deleted from the hash table.
2. drawable B with same address addr is created and added to the hash table
right after 1 is done.
3. st_framebuffers_purge will seach the hash table with drawable addr that
associated with each framebuffer. If drawable is not in the hash table, then
free this framebuffer.

So when drawable B is created, then the framebuffer that associated with
drawable A will not be freed in time. This will cause GTT memory leak.

Since drawable->ID is unique, this creates a hash_table_u64 to store
drawable using drawable-ID.

Signed-off-by: Julia Zhang <Julia.Zhang@amd.com>