brw: disable null_rt unless no output is affected
We found out that some HW changes on Xe2 make the HW avoid reading the
blend state if we're using the null_rt bit in the extended descriptor.

Since the alpha_to_coverage bit resides in the blend state, that state
is ignored and writes are going through to the depth/stencil buffers.

Disable null_rt in the case alpha_to_coverage is enabled or if
depth/stencil/omask are written. Since null_rt exists only on Gfx11+,
currently all driver ensure there is at least one render target setup
for Gfx9, so this is safe to do.

Fixes tests in this pattern on Xe2 :

dEQP-VK.pipeline.*.multisample.alpha_to_coverage_no_color_attachment.*

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Backport-to: 24.2