WIP: iris/gfx12.5+: Keep HIZ_CCS aux usage while sampling from resolved depth surfaces.
This works around graphics corruption seen on MTL platforms when
sampling from a HIZ-CCS depth surface that was previously fast cleared
and resolved for sampling.  Apparently full resolves no longer
guarantee that the CCS surface ends up in a pass-through state due to
the behavior of the L3 cache in presence of compressible data.  In
order to work around the problem this makes sure that we use a
CCS-enabled AUX mode for depth textures if the base surface has a CCS
control surface, even if we are instructed to use ISL_AUX_USAGE_NONE.

This appears to fix the corruption without the need to add extra L3
flushes after resolves (as was done in the Vulkan driver, see
5178ad761c9e8e).

XXX - Switch to partial resolves for better performance since we no
      longer rely on the CCS surface being in pass-through state for
      depth sampling.

v2: Use ISL_AUX_USAGE_HIZ_CCS_WT instead of ISL_AUX_USAGE_HIZ_CCS
    usage to represent the requirements of sampling from a depth
    surface (Nanley).